# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KKN Integration Dashboard API",
        "description": "Dokumentasi API dashboard KKN. API ini dipakai untuk integrasi dengan aplikasi project-akhir (portal mahasiswa & dosen) yang mengirim data pendaftaran, dokumen, dan nilai ke dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Server lokal dashboard"
        }
    ],
    "paths": {
...
        "/api/laporan-akhir/{id}": {
            "get": {
                "tags": [
                    "Laporan Akhir"
                ],
                "summary": "Detail laporan akhir",
                "description": "Mengambil detail satu laporan akhir berdasarkan ID.",
                "operationId": "77dad864fca22688e0dfaeabd2fdb1f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID laporan akhir",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail laporan akhir berhasil diambil"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
        "/api/laporan-akhir/store-from-external": {
            "post": {
                "tags": [
                    "Laporan Akhir"
                ],
                "summary": "Menerima laporan akhir dari project-akhir",
                "description": "Menyimpan laporan akhir beserta file (dikirim sebagai string base64). Endpoint ini juga tersedia sebagai alias di /api/laporan-akhir/store.",
                "operationId": "91953c6fe2d8408dfe39241a99058df9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul_kegiatan",
                                    "user_nim",
                                    "file_content",
                                    "file_name",
                                    "file_mime_type",
                                    "file_size"
                                ],
                                "properties": {
                                    "judul_kegiatan": {
                                        "type": "string",
                                        "example": "KKN Tematik Desa Sukamaju",
                                        "maxLength": 255
                                    },
                                    "user_nim": {
                                        "type": "string",
                                        "example": "2010001",
                                        "maxLength": 20
                                    },
                                    "file_content": {
                                        "description": "Isi file PDF dalam base64",
                                        "type": "string",
                                        "format": "byte"
                                    },
                                    "file_name": {
                                        "type": "string",
                                        "example": "laporan-akhir.pdf"
                                    },
                                    "file_mime_type": {
                                        "type": "string",
                                        "example": "application/pdf"
                                    },
                                    "file_size": {
                                        "type": "integer",
                                        "example": 204800
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Laporan akhir berhasil disimpan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Terjadi kesalahan server"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API key mitra di sini untuk endpoint tarik data nilai.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}